Callbacks and add new value and dataset at runtime

+++ src/Models/Common/Context/HoverContext.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Hover Context
     /// </summary>
-    public readonly record struct HoverContext(double DataX, double DataY)
+    public readonly record struct HoverContext(decimal DataX, decimal DataY)
     {
     }
 }
